Moved category for trace utility network
