+++ src/Android/Xamarin.Android/Samples/Utility network/TraceUtilityNetwork/TraceUtilityNetwork.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2019 Esri.
+// Copyright 2019 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
@@ -7,32 +7,49 @@
 // "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 // language governing permissions and limitations under the License.
 
+using Android.App;
+using Android.Content;
+using Android.OS;
+using Android.Widget;
+using ArcGISRuntime;
 using Esri.ArcGISRuntime.Data;
 using Esri.ArcGISRuntime.Geometry;
 using Esri.ArcGISRuntime.Http;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Symbology;
 using Esri.ArcGISRuntime.UI;
+using Esri.ArcGISRuntime.UI.Controls;
 using Esri.ArcGISRuntime.UtilityNetworks;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
-using Xamarin.Forms;
-using SelectionMode = Esri.ArcGISRuntime.Mapping.SelectionMode;
 
 namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
 {
+    [Activity(ConfigurationChanges = Android.Content.PM.ConfigChanges.Orientation | Android.Content.PM.ConfigChanges.ScreenSize)]
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
-        "Trace a utility network",
-        "Network analysis",
+        "Trace utility network",
+        "Utility network",
         "Discover connected features in a utility network using connected, subnetwork, upstream, and downstream traces.",
         "",
         "Featured")]
+    [ArcGISRuntime.Samples.Shared.Attributes.AndroidLayout("TraceUtilityNetwork.axml")]
     [ArcGISRuntime.Samples.Shared.Attributes.OfflineData()]
-    public partial class TraceUtilityNetwork : ContentPage
+    public class TraceUtilityNetwork : Activity
     {
+        // Hold references to the UI controls.
+        private MapView _myMapView;
+        private RadioButton _addStartButton;
+        private RadioButton _addBarrierButton;
+        private Button _traceButton;
+        private Button _resetButton;
+        private Button _traceTypeButton;
+        private TextView _status;
+        private ProgressBar _progressBar;
+
+        private bool isAddingStart = true;
+
         // Feature service for an electric utility network in Naperville, Illinois.
         private const string FeatureServiceUrl = "https://sampleserver7.arcgisonline.com/arcgis/rest/services/UtilityNetwork/NapervilleElectric/FeatureServer";
 
@@ -44,18 +61,22 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
         private List<UtilityElement> _startingLocations = new List<UtilityElement>();
         private List<UtilityElement> _barriers = new List<UtilityElement>();
         private UtilityTier _mediumVoltageTier;
-        private UtilityTraceType _selectedTraceType = UtilityTraceType.Connected;
+        private UtilityTraceType _utilityTraceType = UtilityTraceType.Connected;
 
         // Task completion source for the user selected terminal.
-        private TaskCompletionSource<string> _terminalCompletionSource = null;
+        private TaskCompletionSource<int> _terminalCompletionSource = null;
 
         // Markers for the utility elements.
         private SimpleMarkerSymbol _startingPointSymbol;
         private SimpleMarkerSymbol _barrierPointSymbol;
 
-        public TraceUtilityNetwork()
+        protected override void OnCreate(Bundle bundle)
         {
-            InitializeComponent();
+            base.OnCreate(bundle);
+
+            Title = "Trace a utility network";
+
+            CreateLayout();
             Initialize();
         }
 
@@ -63,11 +84,11 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
         {
             try
             {
-                BusyIndicator.IsVisible = true;
-                Status.Text = "Loading Utility Network...";
+                _progressBar.Visibility = Android.Views.ViewStates.Visible;
+                _status.Text = "Loading Utility Network...";
 
                 // Setup Map with Feature Layer(s) that contain Utility Network.
-                MyMapView.Map = new Map(Basemap.CreateStreetsNightVector())
+                _myMapView.Map = new Map(Basemap.CreateStreetsNightVector())
                 {
                     InitialViewpoint = _startingViewpoint
                 };
@@ -77,17 +98,17 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
                 UniqueValue mediumVoltageValue = new UniqueValue("N/A", "Medium Voltage", new SimpleLineSymbol(SimpleLineSymbolStyle.Solid, System.Drawing.Color.DarkCyan, 3), 5);
                 UniqueValue lowVoltageValue = new UniqueValue("N/A", "Low Voltage", new SimpleLineSymbol(SimpleLineSymbolStyle.Dash, System.Drawing.Color.DarkCyan, 3), 3);
                 lineLayer.Renderer = new UniqueValueRenderer(new List<string>() { "ASSETGROUP" }, new List<UniqueValue>() { mediumVoltageValue, lowVoltageValue }, "", new SimpleLineSymbol());
-                MyMapView.Map.OperationalLayers.Add(lineLayer);
+                _myMapView.Map.OperationalLayers.Add(lineLayer);
 
                 // Add the layer with electric devices.
                 FeatureLayer electricDevicelayer = new FeatureLayer(new Uri($"{FeatureServiceUrl}/100"));
-                MyMapView.Map.OperationalLayers.Add(electricDevicelayer);
+                _myMapView.Map.OperationalLayers.Add(electricDevicelayer);
 
                 // Set the selection color for features in the map view.
-                MyMapView.SelectionProperties = new SelectionProperties(System.Drawing.Color.Yellow);
+                _myMapView.SelectionProperties = new SelectionProperties(System.Drawing.Color.Yellow);
 
                 // Create and load the utility network.
-                _utilityNetwork = await UtilityNetwork.CreateAsync(new Uri(FeatureServiceUrl), MyMapView.Map);
+                _utilityNetwork = await UtilityNetwork.CreateAsync(new Uri(FeatureServiceUrl), _myMapView.Map);
 
                 // Get the utility tier used for traces in this network. For this data set, the "Medium Voltage Radial" tier from the "ElectricDistribution" domain network is used.
                 UtilityDomainNetwork domainNetwork = _utilityNetwork.Definition.GetDomainNetwork("ElectricDistribution");
@@ -102,44 +123,31 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
 
                 // Create a graphics overlay.
                 GraphicsOverlay graphicsOverlay = new GraphicsOverlay();
-                MyMapView.GraphicsOverlays.Add(graphicsOverlay);
+                _myMapView.GraphicsOverlays.Add(graphicsOverlay);
 
                 // Set the instruction text.
-                Status.Text = "Click on the network lines or points to add a utility element.";
+                _status.Text = "Tap on the network lines or points to add a utility element.";
             }
             catch (Exception ex)
             {
-                Status.Text = "Loading Utility Network failed...";
-                await Application.Current.MainPage.DisplayAlert(ex.GetType().Name, ex.Message, "OK");
+                _status.Text = "Loading Utility Network failed...";
+                CreateDialog(ex.Message, title: ex.GetType().Name);
             }
             finally
             {
-                MainUI.IsEnabled = true;
-                BusyIndicator.IsVisible = false;
+                _progressBar.Visibility = Android.Views.ViewStates.Invisible;
             }
         }
 
-        private void IsAddingStartingLocations_Clicked(object sender, EventArgs e)
-        {
-            IsAddingStartingLocations.IsEnabled = false;
-            IsAddingBarriers.IsEnabled = true;
-        }
-
-        private void IsAddingBarriers_Clicked(object sender, EventArgs e)
-        {
-            IsAddingStartingLocations.IsEnabled = true;
-            IsAddingBarriers.IsEnabled = false;
-        }
-
-        private async void OnGeoViewTapped(object sender, Esri.ArcGISRuntime.Xamarin.Forms.GeoViewInputEventArgs e)
+        private async void GeoViewTapped(object sender, GeoViewInputEventArgs e)
         {
             try
             {
-                BusyIndicator.IsVisible = true;
-                Status.Text = "Identifying trace locations...";
+                _progressBar.Visibility = Android.Views.ViewStates.Visible;
+                _status.Text = "Identifying trace locations...";
 
                 // Identify the feature to be used.
-                IEnumerable<IdentifyLayerResult> identifyResult = await MyMapView.IdentifyLayersAsync(e.Position, 10.0, false);
+                IEnumerable<IdentifyLayerResult> identifyResult = await _myMapView.IdentifyLayersAsync(e.Position, 10.0, false);
                 ArcGISFeature feature = identifyResult?.FirstOrDefault()?.GeoElements?.FirstOrDefault() as ArcGISFeature;
                 if (feature == null) { return; }
 
@@ -154,7 +162,7 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
                     {
                         element.Terminal = await WaitForTerminal(terminals);
                     }
-                    Status.Text = $"Terminal: {element.Terminal?.Name ?? "default"}";
+                    _status.Text = $"Terminal: {element.Terminal?.Name ?? "default"}";
                 }
                 else if (element.NetworkSource.SourceType == UtilityNetworkSourceType.Edge)
                 {
@@ -165,13 +173,13 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
                         double fraction = GeometryEngine.FractionAlong(line, e.Location, -1);
                         if (double.IsNaN(fraction)) { return; }
                         element.FractionAlongEdge = fraction;
-                        Status.Text = $"Fraction along edge: {element.FractionAlongEdge}";
+                        _status.Text = $"Fraction along edge: {element.FractionAlongEdge}";
                     }
                 }
 
                 // Check whether starting location or barrier is added to update the right collection and symbology.
                 Symbol symbol = null;
-                if (IsAddingStartingLocations.IsEnabled != true)
+                if (isAddingStart)
                 {
                     _startingLocations.Add(element);
                     symbol = _startingPointSymbol;
@@ -184,88 +192,72 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
 
                 // Add a graphic for the new utility element.
                 Graphic traceLocationGraphic = new Graphic(feature.Geometry as MapPoint ?? e.Location, symbol);
-                MyMapView.GraphicsOverlays.FirstOrDefault()?.Graphics.Add(traceLocationGraphic);
+                _myMapView.GraphicsOverlays.FirstOrDefault()?.Graphics.Add(traceLocationGraphic);
             }
             catch (Exception ex)
             {
-                Status.Text = "Identifying locations failed.";
-                await Application.Current.MainPage.DisplayAlert(ex.GetType().Name, ex.Message, "OK");
+                _status.Text = "Identifying locations failed.";
+                CreateDialog(ex.Message, ex.GetType().Name);
             }
             finally
             {
-                if (Status.Text.Equals("Identifying trace locations...")) { Status.Text = "Could not identify location."; }
-                BusyIndicator.IsVisible = false;
+                if (_status.Text.Equals("Identifying trace locations...")) { _status.Text = "Could not identify location."; }
+                _progressBar.Visibility = Android.Views.ViewStates.Invisible;
             }
         }
 
         private async Task<UtilityTerminal> WaitForTerminal(IEnumerable<UtilityTerminal> terminals)
         {
-            try
-            {
-                // Switch the UI for the user choosing the junction.
-                MainUI.IsVisible = false;
-                MyMapView.IsVisible = false;
-                StatusGrid.IsVisible = false;
-                PickerUI.IsVisible = true;
-                MyMapView.GeoViewTapped -= OnGeoViewTapped;
+            // Create an array of the terminal names for the UI.
+            string[] terminalNames = terminals.ToList().Select(t => t.Name).ToArray();
 
-                // Load the terminals into the UI.
-                TerminalPicker.ItemsSource = terminals.Select(x => x.Name).ToList();
-                TerminalPicker.SelectedItem = null;
+            // Create UI for terminal selection.
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.SetTitle("Select terminal for junction");
+            builder.SetItems(terminalNames, Choose_Click);
+            builder.SetCancelable(false);
+            builder.Show();
 
-                // Wait for the user to select a terminal.
-                _terminalCompletionSource = new TaskCompletionSource<string>();
-                string selectedName = await _terminalCompletionSource.Task;
-                return terminals.Where(x => x.Name.Equals(selectedName)).FirstOrDefault();
-            }
-            finally
-            {
-                // Make the main UI visible again.
-                MainUI.IsVisible = true;
-                MyMapView.IsVisible = true;
-                StatusGrid.IsVisible = true;
-                PickerUI.IsVisible = false;
-                MyMapView.GeoViewTapped += OnGeoViewTapped;
-            }
+            // Return the selected terminal.
+            _terminalCompletionSource = new TaskCompletionSource<int>();
+            string selectedName = terminalNames[await _terminalCompletionSource.Task];
+            return terminals.Where(x => x.Name.Equals(selectedName)).FirstOrDefault();
         }
 
-        private void Terminal_Selected(object sender, EventArgs e)
+        private void Choose_Click(object sender, DialogClickEventArgs e)
         {
-            _terminalCompletionSource.TrySetResult(TerminalPicker.SelectedItem as string);
+            _terminalCompletionSource.TrySetResult(e.Which);
         }
 
-        private void OnReset(object sender, EventArgs e)
+        private void ChooseTraceType(object sender, EventArgs e)
         {
-            // Reset the UI.
-            Status.Text = "Click on the network lines or points to add a utility element.";
-            BusyIndicator.IsVisible = false;
-
-            // Clear collections of starting locations and barriers.
-            _startingLocations.Clear();
-            _barriers.Clear();
-
-            // Clear the map of any locations, barriers and trace result.
-            MyMapView.GraphicsOverlays.FirstOrDefault()?.Graphics.Clear();
-            MyMapView.Map.OperationalLayers.OfType<FeatureLayer>().ToList().ForEach(layer => layer.ClearSelection());
+            // Create UI for trace type selection.
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.SetTitle("Select trace type");
+            builder.SetItems(Enum.GetNames(typeof(UtilityTraceType)), ChangeTraceType);
+            builder.SetCancelable(true);
+            builder.Show();
         }
 
-        private async void OnTrace(object sender, EventArgs e)
+        private void ChangeTraceType(object sender, DialogClickEventArgs e)
+        {
+            _utilityTraceType = (UtilityTraceType)Enum.GetValues(typeof(UtilityTraceType)).GetValue(e.Which);
+            _traceTypeButton.Text = $"Trace type: {_utilityTraceType}";
+        }
+
+        private async void Trace_Click(object sender, EventArgs e)
         {
             try
             {
-                // Get the selected trace type.
-                UtilityTraceType traceType = _selectedTraceType;
-
                 // Update the UI.
-                MainUI.IsEnabled = false;
-                BusyIndicator.IsVisible = true;
-                Status.Text = $"Running {traceType} trace...";
+                _progressBar.Visibility = Android.Views.ViewStates.Visible;
+                _status.Text = $"Running {_utilityTraceType.ToString().ToLower()} trace...";
 
                 // Clear previous selection from the layers.
-                MyMapView.Map.OperationalLayers.OfType<FeatureLayer>().ToList().ForEach(layer => layer.ClearSelection());
+                _myMapView.Map.OperationalLayers.OfType<FeatureLayer>().ToList().ForEach(layer => layer.ClearSelection());
 
                 // Build trace parameters.
-                UtilityTraceParameters parameters = new UtilityTraceParameters(traceType, _startingLocations);
+                UtilityTraceParameters parameters = new UtilityTraceParameters(_utilityTraceType, _startingLocations);
                 foreach (UtilityElement barrier in _barriers)
                 {
                     parameters.Barriers.Add(barrier);
@@ -281,49 +273,80 @@ namespace ArcGISRuntimeXamarin.Samples.TraceUtilityNetwork
                 // Check if there are any elements in the result.
                 if (elementTraceResult?.Elements?.Count > 0)
                 {
-                    foreach (FeatureLayer layer in MyMapView.Map.OperationalLayers.OfType<FeatureLayer>())
+                    foreach (FeatureLayer layer in _myMapView.Map.OperationalLayers.OfType<FeatureLayer>())
                     {
                         IEnumerable<UtilityElement> elements = elementTraceResult.Elements.Where(element => element.NetworkSource.Name == layer.FeatureTable.TableName);
                         IEnumerable<Feature> features = await _utilityNetwork.GetFeaturesForElementsAsync(elements);
                         layer.SelectFeatures(features);
                     }
                 }
-                Status.Text = "Trace completed.";
+                _status.Text = "Trace completed.";
             }
             catch (Exception ex)
             {
-                Status.Text = "Trace failed...";
+                _status.Text = "Trace failed...";
                 if (ex is ArcGISWebException && ex.Message == null)
                 {
-                    await Application.Current.MainPage.DisplayAlert(ex.GetType().Name, "Trace failed.", "OK");
+                    CreateDialog("Trace failed.", ex.GetType().Name);
                 }
                 else
                 {
-                    await Application.Current.MainPage.DisplayAlert(ex.GetType().Name, ex.Message, "OK");
+                    CreateDialog(ex.Message, ex.GetType().Name);
                 }
             }
             finally
             {
-                MainUI.IsEnabled = true;
-                BusyIndicator.IsVisible = false;
+                _progressBar.Visibility = Android.Views.ViewStates.Invisible;
             }
         }
 
-        private async void TraceTypeButtonPressed(object sender, EventArgs e)
+        private void Reset_Click(object sender, EventArgs e)
         {
-            try
-            {
-                // Prompt the user to select a type of trace.
-                string choice = await ((Page)Parent).DisplayActionSheet("Choose type of trace", "Cancel", null, Enum.GetNames(typeof(UtilityTraceType)));
+            // Reset the UI.
+            _status.Text = "Tap on the network lines or points to add a utility element.";
+            _progressBar.Visibility = Android.Views.ViewStates.Invisible;
 
-                // Set the selected trace type.
-                _selectedTraceType = (UtilityTraceType)Enum.Parse(typeof(UtilityTraceType), choice);
-                TracePickerButton.Text = $"Trace type: {_selectedTraceType}";
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine(ex.Message);
-            }
+            // Clear the utility trace parameters.
+            _startingLocations.Clear();
+            _barriers.Clear();
+
+            // Clear the map layers and graphics.
+            _myMapView.GraphicsOverlays.FirstOrDefault()?.Graphics.Clear();
+            _myMapView.Map.OperationalLayers.OfType<FeatureLayer>().ToList().ForEach(layer => layer.ClearSelection());
+        }
+
+        private void CreateLayout()
+        {
+            // Create a new vertical layout for the app.
+            SetContentView(Resource.Layout.TraceUtilityNetwork);
+
+            _myMapView = FindViewById<MapView>(Resource.Id.MapView);
+
+            _addStartButton = FindViewById<RadioButton>(Resource.Id.addStart);
+            _addBarrierButton = FindViewById<RadioButton>(Resource.Id.addBarrier);
+            _traceButton = FindViewById<Button>(Resource.Id.traceButton);
+            _resetButton = FindViewById<Button>(Resource.Id.resetButton);
+            _traceTypeButton = FindViewById<Button>(Resource.Id.traceTypeButton);
+            _status = FindViewById<TextView>(Resource.Id.statusLabel);
+            _progressBar = FindViewById<ProgressBar>(Resource.Id.progressBar);
+
+            _myMapView.GeoViewTapped += GeoViewTapped;
+            _traceButton.Click += Trace_Click;
+            _resetButton.Click += Reset_Click;
+
+            _traceTypeButton.Click += ChooseTraceType;
+
+            _addStartButton.Click += (s, e) => { isAddingStart = true; };
+            _addBarrierButton.Click += (s, e) => { isAddingStart = false; };
+        }
+
+        private void CreateDialog(string message, string title = null)
+        {
+            // Create a dialog to show message to user.
+            AlertDialog alert = new AlertDialog.Builder(this).Create();
+            if (title != null) alert.SetTitle(title);
+            alert.SetMessage(message);
+            alert.Show();
         }
     }
 }
