+++ src/Android/Xamarin.Android/Samples/Utility network/TraceUtilityNetwork/readme.metadata.json
