+++ src/Forms/Shared/Forms.projitems
@@ -13,7 +13,7 @@
     <None Include="$(MSBuildThisFileDirectory)Samples\Utility network\PerformValveIsolationTrace\PerformValveIsolationTrace.jpg" />
     <None Include="$(MSBuildThisFileDirectory)Samples\Layers\DisplaySubtypeFeatureLayer\DisplaySubtypeFeatureLayer.jpg" />
     <None Include="$(MSBuildThisFileDirectory)Samples\Layers\DisplayAnnotation\DisplayAnnotation.jpg" />
-    <None Include="$(MSBuildThisFileDirectory)Samples\Network analysis\TraceUtilityNetwork\TraceUtilityNetwork.jpg" />
+    <None Include="$(MSBuildThisFileDirectory)Samples\Utility network\TraceUtilityNetwork\TraceUtilityNetwork.jpg" />
     <None Include="$(MSBuildThisFileDirectory)Samples\Network analysis\NavigateRouteRerouting\NavigateRouteRerouting.jpg" />
     <None Include="$(MSBuildThisFileDirectory)Samples\Map\ApplyScheduledUpdates\ApplyScheduledUpdates.jpg" />
     <None Include="$(MSBuildThisFileDirectory)Samples\Network analysis\NavigateRoute\NavigateRoute.jpg" />
@@ -206,7 +206,7 @@
       <DependentUpon>DisplayAnnotation.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Samples\Network analysis\TraceUtilityNetwork\TraceUtilityNetwork.xaml.cs">
+    <Compile Include="$(MSBuildThisFileDirectory)Samples\Utility network\TraceUtilityNetwork\TraceUtilityNetwork.xaml.cs">
       <DependentUpon>TraceUtilityNetwork.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
@@ -839,7 +839,7 @@
     <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Layers\DisplayAnnotation\DisplayAnnotation.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Network analysis\TraceUtilityNetwork\TraceUtilityNetwork.xaml">
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Utility network\TraceUtilityNetwork\TraceUtilityNetwork.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
     <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Network analysis\NavigateRouteRerouting\NavigateRouteRerouting.xaml">
@@ -1473,4 +1473,4 @@
       <Generator>MSBuild:Compile</Generator>
     </EmbeddedResource>
   </ItemGroup>
-</Project>
+</Project>ct>
