+++ src/iOS/Xamarin.iOS/ArcGISRuntime.Xamarin.Samples.iOS.csproj
@@ -272,7 +272,7 @@
     <None Include="Samples\Utility network\PerformValveIsolationTrace\PerformValveIsolationTrace.jpg" />
     <None Include="Samples\Layers\DisplaySubtypeFeatureLayer\DisplaySubtypeFeatureLayer.jpg" />
     <None Include="Samples\Layers\DisplayAnnotation\DisplayAnnotation.jpg" />
-    <None Include="Samples\Network analysis\TraceUtilityNetwork\TraceUtilityNetwork.jpg" />
+    <None Include="Samples\Utility network\TraceUtilityNetwork\TraceUtilityNetwork.jpg" />
     <None Include="Samples\Augmented reality\NavigateAR\NavigateAR.jpg" />
     <None Include="Samples\Augmented reality\DisplayScenesInTabletopAR\DisplayScenesInTabletopAR.jpg" />
     <None Include="Samples\Augmented reality\ExploreScenesInFlyoverAR\ExploreScenesInFlyoverAR.jpg" />
@@ -458,7 +458,7 @@
     <Compile Include="Samples\Utility network\PerformValveIsolationTrace\PerformValveIsolationTrace.cs" />
     <Compile Include="Samples\Layers\DisplaySubtypeFeatureLayer\DisplaySubtypeFeatureLayer.cs" />
     <Compile Include="Samples\Layers\DisplayAnnotation\DisplayAnnotation.cs" />
-    <Compile Include="Samples\Network analysis\TraceUtilityNetwork\TraceUtilityNetwork.cs" />
+    <Compile Include="Samples\Utility network\TraceUtilityNetwork\TraceUtilityNetwork.cs" />
     <Compile Include="Samples\Augmented reality\CollectDataAR\AdjustableLocationDataSource.cs" />
     <Compile Include="Samples\Augmented reality\CollectDataAR\CollectDataAR.cs" />
     <Compile Include="Samples\Augmented reality\NavigateAR\RoutePlanner.cs" />
@@ -743,4 +743,4 @@
   </ItemGroup>
   <Import Project="..\..\ArcGISRuntime.Samples.Shared\ArcGISRuntime.Samples.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\iOS\Xamarin.iOS.CSharp.targets" />
-</Project>
+</Project>t>
